# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Swagger documentation for the OpenCost costmodel HTTP API.",
    "title": "OpenCost API",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "paths": {
...
    "/kapis/costwise.wiztelemetry.io/v1alpha1/assets": {
      "get": {
        "description": "查询集群中的原始或聚合资产数据，包括节点、磁盘、网络、负载均衡器等。\n不传 aggregate 时返回单个 AssetSet；传 aggregate=type 时返回按时间聚合后的资产集合列表。\n参数处理顺序为：先按 filter 过滤资产；如传 aggregate，再按 step 或 accumulate 组织时间桶，最后按 type 聚合。",
        "tags": [
          "Asset"
        ],
        "summary": "查询资产数据",
        "parameters": [
          {
            "description": "时间窗口。必填。支持相对时间和绝对时间范围。示例：window=24h、window=7d、window=today、window=week、window=2026-05-01T00:00:00Z,2026-05-08T00:00:00Z",
            "name": "window",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "按集群过滤资产。支持单个或多个集群，多个集群用英文逗号分隔。该参数会与 filter 做 AND 合并。示例：cluster=prod、cluster=prod,staging",
            "name": "cluster",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "资产过滤条件，使用声明式过滤语法。支持字段：assetType、name、category、cluster、project、provider、providerID、account、service、label[<key>]。常用操作：等于 field:%22value%22，不等于 field!:%22value%22，包含 field~:%22value%22，前缀 field<~:%22prefix%22；AND 使用 +，OR 使用 |。示例：filter=assetType:%22node%22、filter=cluster:%22prod%22%2Bprovider:%22aws%22、filter=label[team]:%22platform%22。注意：URL 中 + 建议编码为 %2B",
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "聚合维度。当前仅支持 type。留空时返回原始资产集合；传 type 时返回按资产类型聚合后的结果。示例：aggregate=type",
            "name": "aggregate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "固定时间桶宽度。仅在传 aggregate=type 时生效。支持 Go duration 风格，如 12h、24h。用于按固定时长切分窗口；不能与 accumulate 同时使用。示例：window=7d&aggregate=type&step=12h",
            "name": "step",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "时间累积方式。仅在传 aggregate=type 时生效。支持：true、all、day、week、month。含义：按自然时间粒度或整个窗口累积；不能与 step 同时使用。示例：accumulate=all 返回整个窗口的汇总；accumulate=day 返回按天累积",
            "name": "accumulate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          }
        }
      }
    },
    "/kapis/costwise.wiztelemetry.io/v1alpha1/assets/carbon": {
      "get": {
        "description": "查询资产的碳足迹估算数据，基于资产成本数据推导碳排放信息。\n该路由仅在启用 Carbon Estimates 时注册；若当前部署未启用，可能返回 404。\n参数处理顺序为：先按 window 取资产，再按 filter 过滤，然后计算碳排放结果。",
        "tags": [
          "Asset"
        ],
        "summary": "查询资产碳排放数据",
        "parameters": [
          {
            "description": "时间窗口。必填。支持相对时间和绝对时间范围。示例：window=24h、window=7d、window=today",
            "name": "window",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "按集群过滤资产。支持单个或多个集群，多个集群用英文逗号分隔。该参数会与 filter 做 AND 合并。示例：cluster=prod、cluster=prod,staging",
            "name": "cluster",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "资产过滤条件，语法与 /assets、/assets/graph 保持一致。支持字段：assetType、name、category、cluster、project、provider、providerID、account、service、label[<key>]。示例：filter=assetType:%22node%22、filter=cluster:%22prod%22%2Bprovider:%22aws%22",
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          }
        }
      }
    },
    "/kapis/costwise.wiztelemetry.io/v1alpha1/assets/graph": {
      "get": {
        "description": "查询资产图表数据，返回按时间分桶后的资产成本曲线。\n参数处理顺序为：先按 filter 过滤资产，再按 step 或 accumulate 组织时间桶，再按 aggregate 聚合，最后按成本降序并应用 offset/limit。\n适合资产趋势图、TopN 图表、按集群/类型/标签观察资产成本变化。\n每个时间片除 items 外，还会返回 totalCost，表示该时间片内所有图表项的总成本，可直接用于计算单项占比。",
        "tags": [
          "Asset"
        ],
        "summary": "查询资产图表数据",
        "parameters": [
          {
            "description": "时间窗口。必填。支持相对时间和绝对时间范围。示例：window=24h、window=7d、window=today、window=week、window=2026-05-01T00:00:00Z,2026-05-08T00:00:00Z",
            "name": "window",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "聚合维度。默认 type。当前仅支持单个资产维度，常用值：type、name、cluster、provider、service、category、account、project、providerID、label:<key>。示例：aggregate=cluster、aggregate=service、aggregate=label:team",
            "name": "aggregate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "固定时间桶宽度。支持 Go duration 风格，如 12h、24h。用于按固定时长切分窗口；不能与 accumulate 同时使用。示例：window=7d&step=12h&aggregate=type",
            "name": "step",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "时间粒度。默认 day。支持：hour、day、week、month。含义：按自然时间粒度返回图表桶；不能与 step 同时使用。示例：accumulate=hour 用于 24 小时趋势；accumulate=week 用于周维度报表",
            "name": "accumulate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "按集群过滤资产。支持单个或多个集群，多个集群用英文逗号分隔。该参数会与 filter 做 AND 合并。示例：cluster=prod、cluster=prod,staging",
            "name": "cluster",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "资产过滤条件，使用声明式过滤语法。支持字段：assetType、name、category、cluster、project、provider、providerID、account、service、label[<key>]。常用操作：等于 field:%22value%22，不等于 field!:%22value%22，包含 field~:%22value%22，前缀 field<~:%22prefix%22；AND 使用 +，OR 使用 |。示例：filter=assetType:%22node%22、filter=provider:%22aws%22%2BassetType:%22disk%22、filter=cluster:%22prod%22%2Blabel[team]:%22platform%22。注意：URL 中 + 建议编码为 %2B，避免被解释为空格",
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "图表项偏移量。默认 0。对每个时间片内按成本降序排列后的结果进行跳过，用于分页或查看 TopN 之后的条目。示例：offset=0 返回前 N 个；offset=10 表示跳过前 10 个",
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "每个时间片返回的最大图表项数量。默认 25。仅影响返回展示数量，不改变底层成本计算。示例：limit=10 返回每个时间片成本最高的 10 个聚合项",
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "costmodel.Response": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "data": {},
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "warning": {
            "type": "string"
          }
        }
      }
    }
  }
}